Remove Restart plasmashell option, replace with desktop file and shortcut
There is no way to do this securely from another user over DBus, and a script with machinectl requires authorisation. This is already a very bad idea for security in the first place.
A compromise is giving the user the shortcut of Meta+Ctrl+Shift+B to restart plasmashell instead.